restore: optimize clang's restore_memcpy implementation
Some local profiling shows this function is a bottleneck, and indeed the
simple optimization here of doing the copy in blocks of qwords instead
of bytes reduces >15% from the total runtime when running with logs
disabled (e.g. make runNoLogs).

While here, add a few expressions to the grep filtering in order to avoid
the need for manually excluding invalid addresses in the restore_memcpy
function.

Change-Id: I2b58dbf5cdbacb6fce175baa2707c17714894c78