tests: add assert coverage reporting
Add assert coverage reporting, which adds two new command-line
parameters to a test binary:

With --log-assert-coverage <file>, each assert (file, line, condition;
test and test case) will be logged to the given file in CSV format.

With --list-asserts the list of asserts in the test binary is listed.

To implement this, statically allocate a struct containing each assert
information, this contains the filename, line number, condition and
(unformatted) failure message.

When an assert is reached, we increment its call-count. As duplicate
assert entries of the same assert may occur, e.g. by an assert in an
inline function declared in a header file and used in different C
files, the later printing will deduplicate using a hash table.

The hash table implemention is https://github.com/troydhanson/uthash
under BSD 1-clause.

The assert code is shared by the controller binary. To prevent dragging
the assert-profiling related code into it, the assert profiling is only
included if the USFSTL_USE_ASSERT_PROFILING define is passed during
compilation. It is enabled for the primary framework and disabled for
the controller.

Also exclude the assert information pointer section from the data-restore
mechanism by changing the appropriate command-line in core.mak.

Change-Id: I66766b8be54b6c6f76abb87802f2fe9b60458a30