multi: decouple controller and participant modes
For some simulation use cases, it may be beneficial to have
a hierarchical set of simulations, i.e.

 simA <-> simB <-> simC

where
 * simA is a multi-controller,
 * simC is a multi-participant, and
 * simB is actually _both_.

To be able to set this up, decouple the controller and participant
modes so that something like simB in the example can be both in a
multi simulation.

This is actually fairly simple, the only thing that really needed
a bit of work is that we treated the local task scheduler as a
participant in the simulation by setting g_usfstl_multi_ctrl_conn
to USFSTL_RPC_LOCAL on the controller, and using the RPC scheduler
linking. This needed to be modified to be direct, not via (local,
fake) RPC.

Change-Id: I7f84f779f6645e053738d690ee0e04e2fbfd33a6