test runner: allow the user to provide an abort handler
Allow the user to provide his own abort handler.
The handler is called when a test aborts as a result of
a failed assert. This is useful for either providing additional
context for a failed assert, or preventing certain asserts from
failing the test under certain conditions, e.g., when running the
test under a fuzzing engine.

Change-Id: Ic85fd0f10fc9d50701e3d55dcc7b2ef8bf9d1a6c