print: improve speed of multi-participant logging
Currently, we're formatting log messages locally, and then
sending them over RPC (async, but still) for actual writing
to the log file. This is obviously rather slow.

In order to improve this, change the scheme so that logger
allocation is still done through RPC in order to be able to
have common log files for different participants, but then
open the real file in append mode ("a"), and directly write
to it. To make that not interleave badly, we need to flush
any buffers before making or returning from RPC calls, but
that's a relatively small price to pay.

In one of our more complex examples, a roughly 1GB log is
being written, and the time on my machine is about
 * 5.6 seconds without logging
 * 69 seconds with the old scheme
 * 38 seconds with the new scheme

This is clearly a significant improvement.

In addition, flushing the logs on RPC calls actually also
fixes an issue with interleaving if loging goes to stdout
instead of a file.

Change-Id: I5b330e75d1ddcd56e6ec1c789f190ffebb52ff1f