sched: add "next_time_changed" callback
In the controller case, we use the scheduler for both the internal
jobs (vhost and network), and for controlling the external apps
(e.g. Linux UML instances.) This causes an issue with free-until,
we don't send it properly to the external applications when adding
an internal job.

For now, to fix this, add a callback to the scheduler that lets us
know when the next time to run at changed - we can then use this
to update the external clients.

Over time, I suspect it might be better to remove this again and
instead link two schedulers together, similarly to the schedctrl
protocol, but inside a single process. However, this is still WIP
and needs some more thought, so add this now to fix the issue.

Change-Id: I4b4f593ae936d941a5c4bb33edebf242b18fe1cf