control: use next_time_changed to fix free-until bugs
As explained in the previous commit, when we have internal changes
to the scheduler we don't tell external applications properly about
a new free-until, causing crashes if they rely on that.

Use the new next_time_changed callback to update the sync points in
these cases. We can then also remove the manual updates since the
addition of jobs for clients also causes this to be called.

Change-Id: I3fe31861038f1be0a266db271fedb3ff5774405f