controller: reset running_client before adding removal job
If we add the removal job before clearing the currently running
client (if the client disconnects that is running), then since
the recent fixes we end up recursing into a free-until update,
which causes us to fail writing a message since the FD has been
closed. This causes us to crash.

Fix this by clearing running_client (if applicable) before the
removal job is added.

Change-Id: I83fed5d418b66c69409bcba83351738072bd2408