fuzz: add --fuzz-repro-list
It may be significantly faster to fork() instead of doing all the new
setup for every time if the test does some significant setup before
calling usfstl_fuzz().

Add a --fuzz-repro-list option that takes as input a file (or stdin)
with a list of files, separated by \n, to use for running. The output
of the forked children is ignored, so this is mostly useful to create
coverage reports.

Also add support for --fuzz-repro-parallel=N, specifying how many
forked children can run in parallel.

Change-Id: I8e0e16cbeaeb3ef211735f8e0c70383fe76d4a23