schedctrl: fix a race with the UM_TIMETRAVEL_WAIT response
As outlined in the comment I'm adding to the code now, there's
a race with the ACK response to UM_TIMETRAVEL_WAIT, because the
message itself passes the "run token" to another process.

Fix this by directly reading the response when the message was
WAIT, instead of polling all the other FDs as well. That will
come later, when we wait for our turn (UM_TIMETRAVEL_RUN.)

Note that in the original Linux commit for extern time travel
mode, commit 88ce64249233 ("um: Implement time-travel=ext"),
I already implemented this the same way, see TTMH_READ use.

Change-Id: Ia6ef8ca3718879a89c977a5b6476a95e5671996c