print: fix logging speed regression on Windows
The speed improvement actually caused a speed regression on
Windows, mostly because we were flushing each log line, rather
than a whole chunk when switching processes, and even in tests
that aren't multi-process were doing that.

Additionally, we were always using "a" instead of "w" mode,
even on the non-multi tests that don't need it.

Fix both of that.

Change-Id: I86eb0d5ef1cc15b24781c55f675d5079e7a0e418
Fixes: 86332452ecbd ("print: improve speed of multi-participant logging")