dwarf: init: iterate only the needed things
If we don't need to resolve static variables or static functions
then there's no point in doing all that work, and if we just have
one of them we don't need to do the other.

Track which of them are needed and resolve only what we want to
save dwarf parsing time.

Change-Id: I95cb166e15f1f2294b8b934f42d172f3c188dc86