dwarf: fix function lookup by name for inlines
When the compiler does some magic such as inlining, we might
get multiple instances of the same function in the debug data,
or similar. For now, we only really want the one that has the
prototype data, since we mostly need it for building the args
as a string for comparisons.

Additionally, we need to add such functions to our vector even
if we don't actually have a PC for them, since they might just
be in the dwarf data for the prototype, due to optimisations.

Change-Id: Iffbda6765dfcc480d2eee9d3085950dc93da5558