override: increase overrides size
If the hash table is less filled, we get better behaviour, so
on relatively large projects it should be bigger. This seems
to be the sweet spot for now for the bigger projects we have,
but we probably need to find a way to tune this based on the
number of functions in the binary or so.

OTOH, having it larger than needed doesn't hurt _that_ much
either.

Change-Id: Ic962c10da0d6abbdb199bc9eb58b949dbae63214