string: optionally allow overriding strcmp/memcpy
If the library is linked to something that's compiled with ASAN
(which is basically the intended use) then the library will link
to ASAN's strcmp() and memcpy() even if it wasn't compiled with
ASAN itself.

Since ASAN's strcmp() is checking everything, it's also quite a
bit slower than normal strcmp() since it accesses both strings
completely, rather than short-circuiting the evaluation.

Since ASAN's memcpy() is checking everything, in clang we were
not able to use it for save/restore since that could touch some
variables that were supposed to be out-of-bounds, and we had a
very slow open-coded memcpy() implementation.

To address both of these, dynamically look up the strcmp() and
memcpy() symbols from the libc directly, for use in the usfstl
framework itself. This is not exposed to the application since
the application should be tested by ASAN.

If the symbol cannot be looked up, fall back to some stupid
implementations of both, just so it works.

Change-Id: I700b757ad29e074ed7f4c536331031595bd5cd45