override: use globals save/restore to reset
A previous commit placed the override system variables in the
norestore section so they wouldn't get saved/restored. However,
this broke the ability to pre-install hooks before the tests
started.

Change it back to allow pre-installing hooks, but instead of
clearing _and_ copying simply remove usfstl_reset_overrides()
instead of relying on it, this even simplifies the code.

Change-Id: I3608ac4ccda3f2bcd8f53e7e90cbefe915e41427