dwarf: add USFSTL_STATIC_TEST_FN_CHK()
Add a checking version of USFSTL_STATIC_TEST_FN() which lets
you not only resolve static functions from the tested module
but also check their types.

This uses the existing machinery of function return/argument
type string creation, and therefore requires giving the args
names in the use of USFSTL_STATIC_TEST_FN_CHK().

Note that in order to avoid compiler warnings/errors, the
return type is changed to a pointer to the original type, so
we can always return NULL in the dummy static inline. This
is compensated at runtime.

Change-Id: Ia0a1c0760a6e0ef7c50a4596823355b84b96c4cb