dwarf: optimise function lookups for static function references
We currently do things rather stupidly: first, we iterate all the
dwarf functions. Then, if we find one that we need a reference for,
we look up its data again - by name. Pass the necessary information
to _directly_ get the data for the one found during iteration, so
we can avoid the second lookup.

Additionally, this lets us really find all instances of a given
function through the iterator.

To simplify the API, add backlinks from the function to the unit
and from the unit to the dwarf_data they belong to.

Change-Id: If56fc00174044a1de60cbd219f313ff9a2dc0f09