dwarf: optimise static function resolving
We're currently iterating *all* functions, and then for each
iterating the resolve list. However, typically, the resolve
list has two to three orders of magnitude fewer entries than
the total list of functions.

Optimise this a bit by adding the ability to not iterate all
functions but rather only all dwarf entries matching a given
function name as a filter, so we can bsearch() for the right
name first and then iterate from there up and down.

Change-Id: I33a0bbf4b62af8f8b57702193778fdcbbec6b309