main: exit with failure status if marked-as-failing test passes
For tests that are marked as failing, reverse the logic of when to
return a non-0 exit status: if it failed (as expected) return 0, and if
it passed return non-0.
This could help in finding tests that were marked as failing but are
now passing so would presumably stay that way.

Change-Id: Ie1f27e9e22a327b2d9b6225dc2d5a8d9c9a3b61e