build: prefix the tested-lib target with "tested-"
In case both the tested and support build targets in the user's Makefile
use the same build configuration system, make could get confused by the
all-wildcard rule of the tested configuration target when figuring out
what the configuration rule of the support is.
Solve this by forcing a "tested- prefix to the tested-lib target. This
also makes things more regular and unifrom - the support lib is
placed in $(USFSTL_BIN_PATH)/support-%, and target lib is placed in
$(USFSTL_BIN_PATH)/tested-%.

Change-Id: I772ed039591d538377bf1f2e55aaf236b2be07aa