control: fix another startup race condition
As is obvious in hindsight, we have three states that
clients can be in:
 * connected, and thus added to the usfstl loop list
 * having requested start, but not gotten an answer yet
 * having started fully

The first two of these were mixed up in the code since
I only had a "bool started", and thus the following
scenario occasionally happened:

 - client 1 connects, is added to the list and sends
   a START message
 - client 2 connects and is added to the list
 - we process the list
   - client1: we send an ACK, wait for it to request
              WAIT, ACK it and continue
   - client2: we send an ACK (with wrong seqno) and
              wait for it to request WAIT

At this point we wait forever, because client 2 is
itself waiting for us to answer its START message
with an ACK first, but we think we've already done
exactly that.

Fix this by actually separating client state into the
three appropriate states (CONNECTED, START_REQUESTED,
STARTED) so that we only process those clients on the
list that actually sent their START message already.

Change-Id: I8c93a43162edac5c1d7735f47b3374525d4c917f