vlab: add a workaround for gdb module breakpoint issues
As explained in the comment, gdb has issues with userspace clone()
vs. module breakpoints due to the MAP_SHARED nature of module code.
Work around that by installing a breakpoint (that does nothing but
continue) on init_new_ldt() which is just after the clone().

Also add a nice welcome message that warns users not to remove that
breakpoint.

Change-Id: I70a515e346b85a50203941231a5d07e2ea83c193