alloc: check for valid pointer before realloc/free
If we have an invalid pointer it's nicer to detect that and
abort the test in a controlled manner, saying what the bug
was, rather than potentially crashing the underlying malloc
implementation.

Change usfstl_realloc() and usfstl_free() to do that by
calling usfstl_alloc_remove(), which checks this, first.

Change-Id: I639bc5b6009e1986e0455339e67b0ddeb451b3a4