[BUGFIX] rpc: fix log flushing in multi simulation
The call to usfstl_flush_all() was misplaced, it was in the
*receive* side of a call, i.e. when somebody else is making
a call to us. Flushing at this point is pointless, there's
nothing in the buffers. Instead, flush when we're *making*
a remote call, so that our buffers are flushed before the
other side invokes any code that might print anything.

Fixes: 86332452ecbd ("print: improve speed of multi-participant logging")
Change-Id: I002213c8896886e69dffb29800e28db0199d1124