build: separate USFSTL_FUZZING=repro build
With USFSTL_FUZZING=repro we shouldn't build in the normal
'usfstl' folder, but rather in a separate one, move the
line that changes the target to fix that.

Without this, we saw build issues such as building first
without fuzzing and with USFSTL_FUZZING=repro not doing a
correct build.

Fixes: fb26b42cb862 ("build: separate USFSTL_FUZZING=repro build")
Change-Id: I8eaf2ae310870b3cbe4032937e30b254291f1cdd