multi: fix participant startup
As it turns out, the participant startup is broken - they're marked
as having started immediately when they call the test_started() RPC,
but then they actually start the test and start operating, and so
does the controller, so they're running in parallel, which is clearly
wrong.

Instead, wait for them all to have entered WAIT state, so that then
the controller can actually run.

Change-Id: I149549a1b50f5daa24fa18b486f84c64fe2f3938