multi: exclusively use multi-scheduler hierarchy
Always link the task scheduler to the multi scheduler using
the normal usfstl_sched_link(), without even overriding the
behaviour, and completely use the multi-scheduler (which can
be in a hierarchy, if multi-layer simulations are done) for
the multi-participant simulation.

This also requires moving the shared memory updates to the
right places where waiting happens.

Change-Id: I02b02211da3a092d76f8e46011a0ec128086cae9