sched: fix assertion in usfstl_sched_set_time()
Originally, usfstl_sched_set_time() was used only for external
requests, but when converting the multi-participant simulation
in the next patch we need to also use it in other scenarios.

Change usfstl_sched_set_time() to just assert on the current
first job, not on the requested time, since that's still all
good.

Change-Id: Ida1c203e77ee784027f4abad89b4eeb76868ba66