multi: don't use async RPC
Async RPC is significantly broken, since it breaks the
underlying "single-threadedness" of the entire thing,
and that wasn't taken into account at all, and is very
difficult to take into account.

As such, don't use async RPC but instead complete the
calls that used to be async later in the flow. This is
then after the function actually returns, but as the
caller is still waiting for it to "return", this then
requires a new function in the RPC framework that can
be called there, named usfstl_rpc_send_void_response().

Change-Id: I0e9a70a15ab312242f75390fe2763300ae2d019b