multi: fix test startup and simplify waiting
During startup, we currently start tests on all participants
and then wait for all of them to enter WAITING state (i.e.
wanting to be scheduled). This is wrong, as it creates some
concurrency: all of the startup in the participants happens
in parallel. That wouldn't be so much of an issue, but then
they can all issue RPC calls to us at the same time, and in
turn we might issue RPC calls to the next layer up, which
then causes races since the RPC code isn't really made for
concurrency.

In order to avoid that, wait for each participant to start
scheduling and waiting to be scheduled after starting the
test.

Also, when we actually ask some participant to run, we only
need to wait for that one to finish and enter WAITING state
again, so we can significantly simplify the wait code and
replace usfstl_multi_controller_wait_all() by just waiting
for a single participant: usfstl_multi_ctl_wait().

Change-Id: Ia7a1dbb14709e61298771bc060ff790e4accc91e