dwarf/rpc: support cross-rpc backtraces
When something fails during an RPC call, the whole thing
gets really difficult to debug. In order to simplify it
somewhat, support backtracing across RPC, i.e. jump over
to the RPC caller when we encounter an RPC call on the
callee, and then print on the caller only the relevant
segment, after continuing on the callee. This may be in
a recursive fashion too, if recursive RPC calls were
made in the system.

Note that we complete the backtraces on all participants
even if the logical flow completes alread at the top of
the RPC hierarchy, this is marked with "finish backtrace
on ...".

Change-Id: I1183ae8c07a46e30bfebae1a21baf396bd4bcfd1