multi: allow configurable scheduler and expose RPC connections
Introduce usfstl_multi_add_rpc_connection_sched() to add an RPC
connection to the multi-participant framework, allowing peer-to-
peer communication that synchronizes the given scheduler. If it
is given as NULL, then synchronizes the g_usfstl_top_scheduler.

Expose both this and usfstl_multi_add_rpc_connection() to allow
setting up the peer-to-peer RPC connections using it.

Change-Id: I4e28089339e232b92ddcee205452ef5518bdbdb7