rpc: allow handling disconnects
In some cases, we want to handle/tolerate disconnects on an RPC
socket instead of just crashing on the next read, e.g. when using
the new usfstl_multi_add_rpc_connection_sched() API we might want
to also remove the connection again.

Handle this by making an RPC call when usfstl_rpc_del_connection()
is called that removes the connection on the other side - if it's
in fact set up to tolerate disconnects, otherwise continue as we
do now which will eventually crash the system (but that'll be the
desired behaviour if something unexpected removes the connection.)

For internal use in multi simulations, where we call an RPC that
causes the binary to exit immediately after returning a response,
add and use usfstl_rpc_del_connection_raw().

Change-Id: I8487019feed30712c3ce78d77145621d0b4a654c