[BUGFIX] sched: fix sync time calculation
If external sync time was set in the past but hasn't been
updated, then when we go to propagate the sync time to
further downstream multi-test participants, we might send
them a sync time that's actually older than the current
time, which makes no sense and causes them to assert. Fix
this by calculating using the sync time only if it's in
the future.

Change-Id: I99a21db618c346eb38242808d2585c4e921a82f1