schedctrl: add usfstl_sched_ctrl_yield()
As described in the documentation, usfstl_sched_ctrl_yield()
allows yielding to other tasks that may be runnable in the
controller application, via the connection to the controller.
It's implemented simply by rescheduling itself and then just
waiting for getting that new scheduling round, this works as
the controller will always put new requests last.

Change-Id: I0ef1495f8867f7723e2ca0864cb6198d51afcaa5