override: realign stack at usfstl_find_repl's entry
The stack alignment requirements have changed in the X86-32 ABI,
requiring the stack to be aligned to 16-bytes before each function call.

fentry does not adhere to this requirement when calling
usfstl_find_repl, and the stack pointer ends up being 8 modulo 16.

Aligning the stack in fentry is a tad tricky because of how
usfstl_find_repl works, so align the stack on usfstl_find_repl's entry
instead.

Change-Id: I2c52f1971c743c816c69b6f607d119fde7431204