sched/task: allow blocking individual jobs/tasks
We currently have a mechanism to block groups of jobs and
tasks, through a bitmap, but no way to independently block
execution of a specific job. This may be needed however to
implement some primitives the OS (such as ThreadX) might
have in the real system.

Add a mechanism to block an individual job and extend the
task API to allow doing that as well.

Change-Id: Iba30cb609f11148ae55d5097b472d050985dc73a