vlab: initialize random better
So ... haveged has a bug, fixed here:
https://github.com/jirka-h/haveged/pull/70/

which causes it to hang in an infinite loop under
certain conditions.

Due to the pool size decreasing from 4096 to 256,
this is easier/faster to hit in 5.18-rc1 and later
kernels.

But we really don't care about any of this, so
just pretend we have some randomness.

Change-Id: Ia5d0541cab984a831404dfb05e86d14247403057