task: fix UAF when tasks exit
We currently have this block of code in the task code:

	do {
		ran = usfstl_sched_next(&g_usfstl_task_scheduler);
	} while (!usfstl_task_from_job(ran));

this is meant to continue there if some non-task jobs
were selected by the scheduler and ran.

However, we could have a situation where multiple tasks
are scheduled. I'm showing the function calls under the
*stack* that they happen, not which task they *logically*
belong to:

task1/job1             task2/job2
 // suspends
 usfstl_task_next()
  usfstl_sched_next()
   //selects job2
   usfstl_task_job_fn()
    usfstl_switch_ctx(task2->ctx)

                        usfstl_task_ctx_fn()
                         task->fn() // returns
                         usfstl_task_end_self()
                          g_usfstl_task_to_end = task2
                          usfstl_task_next()
                           usfstl_sched_next()
                            //selects job1
                            usfstl_task_job_fn()
                             usfstl_end_self(task1->ctx)

    // continue as if returned from
    // usfstl_switch_ctx(task2->ctx)
    // task2 has been destroyed
   // usfstl_task_job_fn() returns
  // usfstl_sched_next() returns job2!

  // and this (from the above code snippet) becomes
  // use-after-free since job2 has been freed with
  // the task2 and its context
  usfstl_task_from_job(ran) // CRASH

To fix this, use a global variable, g_usfstl_task_selected,
that indicates whether or not the scheduler selected a task.
It's set to false immediately before scheduling and to true
when usfstl_task_job_fn() returns to usfstl_sched_next(),
which then returns to the caller since a job was run. Thus,
we know out-of-band whether or not a task executed, and can
avoid looking at the returned job pointer.

Change-Id: I31c9582f0052ef0111abe6e9584ed0ea0d7ac8b1
Reported-by: Michael Rogov Papernov <michaelx.rogov.papernov@intel.com>