[BUGFIX] controller: don't send messages to dying clients
If a client is dying (we've called remove_client()), then
don't try to send it any more messages, which would cause
us to call remove_client() again, which crashes.

This happens e.g. if the write in update_sync() fails and
we remove, but then we try to send the RUN message anyway
in run_client().

Fixes: b14716c66b45 ("[BUGFIX] controller: don't send messages to dying clients")
Change-Id: Ica6755b27c346d0533f924e69668f605f60dd30a