test: fix integer size cast warnings
On 64-bit, casting a pointer to int results in a warning
("cast from pointer to integer of different size"). Fix
that by using uintptr_t as the integer type here.

If the offset were too big for unsigned int (which really
hopefully doesn't happen - who has structures that big?)
we should get an integer truncation warning on the init.

Change-Id: I524920736c5e84ca2e76db222ce0f3e1c232421e