multi-ctl: add an option to disable internal sync updates
This option is utilized when the multi-ctl is managed by an external
scheduler process, and any synchronized updates from external scheduler
are not transferred to the multi-ctl scheduler.

It enables that if participants (multi-ptc) want to work with multi-ctl,
they always need to request time from the controller to run (no sync
messages are sent to them).

It should be noted that using multi-ctl with an external controller
is not optimal, and the recommended approach in such cases is to
bypass multi-ctl altogether and connect the multi-ptc directly to
the external controller. However, if using multi-ctl is unavoidable,
then this option can be used to manually trigger synchronization updates
between the multi-ctl and the external scheduler process.

Change-Id: Ib5aa1225da872a83637c149478064f5a70a294d4