examples: common.lib: change assumption of first machine (m1) name
We change the assumption that first machine name is only m1 and now
allow it to have extension m1-XYZ or m1_XYZ.

This feature is helpful in cases where we want to add a description to
the machine that indicates its purpose.
It allows us to quickly identify and find logs of failed tests,
improving our debugging process.

Change-Id: I791e8448dd1cc2e4d807591e281aa82095fb5f1c