testrun: prevent possible infinite abort loop
It is possible to hit an assert on post-test which will call
usfstl_abort(), then we'll longjmp and hit the same assert
on post test again and so on in infinite loop.

Prevent this by adding another testcase_status which will be
set in case usfstl_abort() is called twice and then skips the
test->post() call.

Change-Id: I081df16d4dae4493e4fdeca28fb0666b709c5fdb