vlab: Use a pristine root directory with defined host mappings
Create a pristine root directory and bind mount everything from hostfs
as desired. By default, all directories in / that do not have a default
definition will be bind mounted to the new root directory. However, the
user can also configure the precise set of mounts that will be done.

This removes the old rootfs option, it should be simple to replace it
with host_binds mappings.

Change-Id: I61b9eb2dbb9c61bc2bf4f6b732e5205ce4423e67