loop: simplify select() handling
We don't need the exception set, since we anyway ignore
it. This was a consequence of me having a bug, and then
Naftali fixing it without realising the exception set
was then irrelevant. Remove it.

Then we also don't need the outer loop since select()
can only return with something set (since we assume no
signals for now, as encoded in the assert.)

Change-Id: I0126548646c42b33572dfedf21b8101dd02f9bd9