[BUGFIX] vlab: ensure we wait for controller node first
We assume that everything is shut down (or at least almost done shutting
down) when the controller node finishes. Other processes may quit
much earlier when drivers are unloaded for example. As such, we need to
wait for the controller node to quit in order to know that the vlab run
is finished.

Fixes: 86001d49e57f ("vlab: start node plugins separately")
Change-Id: If45fd0cff945f4822289222610c62f6b9c64d925