vlab: use processes for start_process()
The simulation really should shut down gracefully,
and we've tried to ensure that for our setup. So
make the default here to expect graceful shutdown
instead of killing, and exceptions can use the old
way of starting.

However, then we need to not use this for wmediumd
because it won't shut down by itself, and then it
keeps the controller alive as well.

Change-Id: I2228803074bcb58d76f9365c8c57e9c7ee0e04bf