task: fix semaphore wakeup sequencing issue
During debugging of one of our simulations, we found that
(somehow only sometimes, which is a separate issue) it seems
to have created a scenario like this:

Task A:
  usfstl_sem_wait(S1);
Task B:
  usfstl_sem_post(S1);
  usfstl_sem_wait(S1);
  usfstl_sem_wait(S2); // or something else to suspend itself

In this case, the code will currently signal A to wake up and
increment the counter of S1 for A to consume it. However, then
B's usfstl_sem_wait() immediately consumes the counter, and
when A is actually started as B suspends itself, it finds the
counter 0, even though no timeout happened, and asserts.

Semantically, the wait queue of the semaphore should be that,
a queue, so when A is woken up it really should be the one to
decrement the counter.
FIXME: what about priorities?

Implement that by never even incrementing the counter when a
task is woken up and just marking that task as owning it now.
This ensures that B will wait for S1 and A will get it.

Change-Id: Ib5542be208abb2754fb71ac51f30d2f140ddfec8