vlab: gdb: reinstall breakpoints after start_userspace
The init_new_ldt function disappeared upstream and we therefore cannot
use it to hook after the clone() anymore. Change the code to install
breakpoint on start_userspace, then finish the function using a second
breakpoint and continue from there.

Doing this will work with newer (and older) UML kernels and the change
also means that the breakpoint is now internal and invisible to the
user.

Change-Id: Id084b71ba17512fac744c21fcf36c7e9479d3e2d