vhost: do not map memory below mmap_offset
We had some odd thing where we also mapped everything below mmap_offset.
It appears that this was original copied from elsewhere and it was done
there to enable the use of huge pages.

In our case, we are usually mapping UML physical memory which will have
a large hole at the beginning to account for the memory of the loaded
executable. This causes the mapping to become much larger than needed
and can cause issues if a 32bit process is handling multiple devices.

Change-Id: I031bab9ff3dcf5369813aadecc965f3d379d0eaf