sharedmem: local view of simulation not updated fix
in cae of sched_wait is called from other simulation, with a shared
memory section update, the local view of the called simulation not
updated.
later when the called simulation generating a shared memory update
to other simulation, the local view is not sync, and thus reaching
logical assert condition.

Change-Id: I4ff89e65f72a9a9c49939cf70560fdde06ade494