sharedmem: merge only relevant sections
the merging of the sections need to be performed only on relevant
sections. the relevant sections are determined using the actual
sections that are compiled as part of the current simulation,
and not as part of the recieved message of memory update.
thus in case where recieved message of not relevant memory section,
or actual compare of the memory is not different from the local version
of it, no reason to mark as the memory outdated for later maintainance

Change-Id: I7ac79be0e75cc12e84c1829b25c82fa10e216855